Huge commit, reduce coplexity in handler and discovery, replace apache with jetty, using newer jetty unitl ESH upgrades
