+++ addons/binding/org.openhab.binding.venstarthermostat/.classpath
@@ -7,8 +7,10 @@
 		</accessrules>
 	</classpathentry>
 	<classpathentry kind="src" path="src/main/java"/>
-	<classpathentry kind="lib" path="lib/httpclient-4.3.4.jar"/>
-	<classpathentry kind="lib" path="lib/httpcore-4.3.3.jar"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry kind="lib" path="lib/jetty-client-9.4.10.v20180503.jar"/>
+	<classpathentry kind="lib" path="lib/jetty-http-9.4.10.v20180503.jar"/>
+	<classpathentry kind="lib" path="lib/jetty-util-9.4.10.v20180503.jar"/>
+	<classpathentry kind="lib" path="lib/jetty-io-9.4.10.v20180503.jar"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
