--- addons/binding/org.openhab.binding.venstarthermostat/OSGI-INF/org.openhab.binding.venstarthermostat.internal.discovery.VenstarThermostatDiscoveryService.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.2.0" configuration-pid="binding.venstarthermostat" immediate="true" name="org.openhab.binding.venstarthermostat.internal.discovery.VenstarThermostatDiscoveryService">
-   <service>
-      <provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
-   </service>
-   <implementation class="org.openhab.binding.venstarthermostat.internal.discovery.VenstarThermostatDiscoveryService"/>
-</scr:component>
